# download/openapi-ad-service.yaml
openapi: 3.0.3
info:
  title: CargoBit Ad Service API
  description: |
    Produktionsreife API für Werbeanzeigen auf der CargoBit-Plattform.
    
    ## Features
    - Banner-Ad-Platzierung (970×250, 300×600, 300×250)
    - Real-Time Targeting (Route, Risk-Level, User)
    - Impression & Click Tracking
    - Kampagnen-Verwaltung für Partner
    
    ## Monetarisierung
    - CPM (Cost per 1000 Impressions)
    - CPC (Cost per Click)
    
    ## Authentifizierung
    Lesende Endpunkte sind öffentlich. Schreibende Endpunkte erfordern Authentifizierung.
  version: 1.0.0
  contact:
    name: CargoBit API Support
    email: api@cargobit.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.cargobit.io/v1
    description: Production
  - url: https://api-staging.cargobit.io/v1
    description: Staging
...
paths:
...
  /ads/click:
    post:
      tags:
        - Tracking
      summary: Klick tracken
      description: Registriert einen Klick für Analytics und Abrechnung
      operationId: trackClick
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClickRequest'
      responses:
        '200':
          description: Klick erfasst
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClickResponse'
...
components:
  schemas:
...
    ClickRequest:
      type: object
      required:
        - impressionId
        - adId
      properties:
        impressionId:
          type: string
        adId:
          type: string
        timestamp:
          type: integer
        clickPosition:
          type: object
          properties:
            x:
              type: integer
            y:
              type: integer
    ClickResponse:
      type: object
      properties:
        success:
          type: boolean
        click:
          type: object
          properties:
            clickId:
              type: string
            impressionId:
              type: string
            adId:
              type: string
            recorded:
              type: boolean
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT