# src/app/api/ads/click/route.ts
import { NextRequest, NextResponse } from 'next/server';

/**
 * POST /api/ads/click
 * 
 * Track an ad click
 * 
 * Request Body:
 * - impressionId: string (required) - Impression ID from the ad view
 * - adId: string (required) - Ad identifier
 * - timestamp: number (optional) - Unix timestamp
 * - clickPosition: object (optional) - { x, y } click coordinates
 */
export async function POST(request: NextRequest) {
  try {
    const body = await request.json();
    
    const { impressionId, adId, timestamp, clickPosition } = body;

    // Validation
    if (!impressionId || !adId) {
      return NextResponse.json(
        {
          error: 'ValidationError',
          message: 'Impression-ID und Ad-ID erforderlich',
          code: 'MISSING_REQUIRED_FIELDS',
        },
        { status: 400 }
      );
    }

    // In production:
    // 1. Validate impression ID exists and matches adId
    // 2. Check for click fraud (duplicate clicks, bot detection)
    // 3. Store click in analytics database
    // 4. Deduct from campaign budget if CPC model
    // 5. Trigger conversion tracking pixels
    // 6. Update real-time dashboards

    // Mock response
    const click = {
      clickId: `clk_${Date.now()}_${Math.random().toString(36).substr(2, 9)}`,
      impressionId,
      adId,
      recorded: true,
      timestamp: timestamp || Date.now(),
      clickPosition: clickPosition || null,
    };

    return NextResponse.json({
      success: true,
      click,
    });

  } catch (error) {
    console.error('Click tracking error:', error);
    return NextResponse.json(
      {
        error: 'InternalServerError',
        message: 'Fehler beim Tracking des Klicks',
        code: 'CLICK_TRACKING_FAILED',
      },
      { status: 500 }
    );
  }
}